CBNGL: Replaced MCell's "+" with BNGL's "," for accumulating counts.
This change looks for strings like "WORLD]+COUNT" (with optional spaces)
and converts them to "WORLD],COUNT" before any further processing. This
is intended to handle summation of count statements (subtraction isn't
addressed).

Note that this operation might be combined with subsequent WORLD/COUNT
processing, but it was done as a separate step for clarity.