CBNGL: Added "_MDLString" to file names as a CBNGL Engine option.
The CellBlender convention is to add "_MDLString" to any plots that are
created with an MDL string. This string wasn't being added by the CBNGL
engine when post-processing. The files could still be plotted, but they
would have to be found and loaded manually (if the plotter supported
manual loading).

The use of an "MDLString" for plotting BNGL molecules was a quick way to
get plotting functionality without a lot of code rewriting. At some
point CellBlender should begin to handle BNGL syntax more naturally.