First attempt at timed creation of a CellBlender Tutorial Model
This commit attempts to create the Organelle model in a step by step
manner where each step is reflected in Blender. Eventually, this would
allow capturing each frame of a tutorial automatically.

The approach taken here builds on the timer event code that we already
use to update the progress and scroll the text output of a run. This
code is a little more complicated because it must also build a model.
Additionally, it would be good to have the same model be built either
via the step by step process or in a single step (current test suite
mode of operation).

To make this work, each step is matched against a current step number
before it is executed. The step numbers are incremented by the timer
callback, and the build process is called again with the new step number
in place. Only the currently active step number will be executed. Each
step is also gated with a -1 to allow for single pass creation of the
entire model.

Note that the earlier model construction scripts used a lot of local
variables. However, for the script to operate across multiple passes,
these local data items must be stored. The "self" object is
appropriately used for that purpose.

For some reason, this version gives erratic behavior, and the timer
doesn't seem to stop (even though the cancel is called). This is still
a very preliminary version.