Changed default for CBNGL's postprocess to use "_MDLString" file name.
The normal way to specify BNGL reaction data output is through the MDL
String mechanism. So CellBlender will look for files with _MDLString as
part of the file name. The CBNGL engine can either add that string or
not (based on a user setting). It had been defaulted to NOT add the
string, but that will generally not work. This commit changes the
default to add the _MDLString prefix when postprocessing the data.