Accept single colon in release site expressions.
MCellR has been requiring a double colon to specify compartments in
release sites. The bngl_to_data_model.py code was modeled after that
behavior. However, it has come to light that the proper syntax is a
single colon.

This commit changes the bngl_to_data_model code to only accept the
single colon. This will not work with the current MCellR, and the single
colons will need to be translated to double colons manually in each
release site to work with the current MCellR. For that reason, this
change was made in a different branch until either the exporting can
export a double colon or MCellR can be fixed to accept a single colon.