Modified CBNGL to data_model code to ignore text after "end model"
The CBNGL to data_model code (bng/bngl_to_data_model.py) had been
written to expect files that started with "begin model" and ended with
"end model". The additional command lines were not expected. This commit
modifies that file to ignore any lines past the "end model" line.

Note that this was done by re-processing the entire list of lines for
clarity (and simplicity). If we find this to be a performance
bottleneck, that processing could be done more efficiently.