Added quick fix to import compartment data from SBML files (.xml).
The cellblender/bng/__init__.py file has not imported any compartment
data in the past.

This quick fix pulls the compartment data from the .xml file and
generates a section of BNGL code representing the compartments. It then
calls the BNGL importing code (from bng/bngl_to_data_model.py) to build
a data model. Then the appropriate CellBlender calls are made to rebuild
CellBlender's data from those parts of the data model.

This seems to work, but for some reason, molecules that don't exist at
the start of the simulation do not seem to be properly initialized. This
may be a separate issue and may require another commit.