Updated the bng/external_operators.py file to newer molecule creation
The SBML importer has gotten out of date by at least 2 years. The
molecule creation methods have changed in that time, so the code that
was creating molecules from BNGL was not up to date. In particular, it
was not properly creating the molecule templates in Blender. This commit
updates the bng/external_operators.py to use the new creation method.

Note that there may be other parts of the BNGL importing code that are
similarly out of date. This commit only addresses the very narrow issue
of creating new molecules (and may not even address that completely).