Force "NULL" as product of SBML imports with no listOfProducts section
The SBML importer did not set any value for products of reactions with
no listOfProducts section in the XML file. MCell requires an explicit
"NULL" string to be inserted if there are no products.

This commit checks all reactions as they are imported for empty
"products" strings and replaces them with NULL.

As noted in the previous commit, there may be many other parts of the
SBML importer that do not function properly.