Updated makefile's "id" target to show real path (needs "realpath")
Now that the makefile is installing to ~/my_cellblender_link/ it can be
more difficult to know exactly where it's putting the installation. The
"id" target (make id) has shown the CellBlender ID in the past, and this
change just adds a few lines to use the "realpath" command to translate
any links to their actual paths. This was added to the "id" target and
not to the "install" target because the "realpath" command might not be
available on all platforms. Even though that would have been the last
thing done (no consequences for failure), it's better to have the actual
install not produce any errors.