Use full path of MDL file when writing a variable rate constant file.
CellBlender makes use of a few Python modules that were originally
written to operate in different contexts. The "data_model_to_mdl" family
of functions fall into that category. In some contexts, these functions
are called with the expectation that their working directory is the MDL
file location (where new MDL is written). In other contexts, that may
not be true.

This is generally not a problem since most of these functions accept a
file handle to an already opened file. They simply write into that file
handle without knowing anything about the path to that file.

The only exception is the "write_reactions" function which may sometimes
need to create a variable rate constant file given only a file name (no
path). If the program is being run from the location of the MDL files,
then this will work fine (as it does in some contexts). However, if it
is being run from some other directory, then it needs to find the proper
path to create the variable rate constant file. The previous version
wasn't doing that, and it was creating the variable rate constant file
in the current working directory of the process rather than in the
proper location relative to the MDL.

This commit creates the new variable rate constant file at the same
location as the MDL file that it's currently writing. This is the "f"
parameter passed into the "write_reactions(reacts,f)" function.