Added "x", "y", "z", and "a" spatial location to components.
Each molecule component can now have an x,y,z,a to specify its location
relative to the molecule itself. These are "panel parameters" in
CellBlender, but they are currently being displayed in a template list.

Panel parameters have traditionally been drawn in a panel (and not in a
template list). As a result, the "draw" function for panel parameters
has always drawn the name, info button, and other optional items. These
items don't fit well in a single row template list. A more compact
drawing mechanism was therefore needed.

This commit adds a new function named "draw_prop_only" to the parameters
to only draw only the parameter's property itself (the expression string)
and not any of the other attributes (name, info button, etc).

This commit also adds the fields: "x", "y", "z", "a" to each component
in the "bngl_component_list" of the data model.