Added support for an electric species via the User Interface Panel
This commit adds a text field to the Prototype_Andreas_1 run panel for
entering the molecule names of any molecules that will be created as
members of the Electric Field class.

Molecule names should be entered as comma-separated values with no
spaces because they will be passed as a command line argument to
run_sim.py.

This commit demonstrates how simulation engines can add user interface
items that are not otherwise supported by CellBlender. CellBlender has
no field to mark molecules as "diffusable" via an electric field. But
using the custom fields in an engine panel, these values can still be
passed to a simulation engine.

Note that values stored in these panels are not currently stored in
either the .blend file or the data model.