Improved the delete_molecules algorithm in Proto_Andreas_1
The previous code always deleted molecules from the end of the list.
This would be fine if the molecules were created equally (in a single
release). But if the molecules were released differently (different
locations or times), then those released last would disappear first.

This commit changes the simple deletion code of:

    self.molecule_list = self.molecule_list[0:num_left]

to a slightly more complex (but more appropriate) version:

    self.molecule_list = random.sample(self.molecule_list, num_left)

This is just a toy prototype, so there's no real change to CellBlender.