Added a function to dump a mol-viz file (must be called manually).
During the development of the spatially structured molecules, an attempt
was made to include components as molecules for visualization. This
created extremely large visualization files, and a mol_viz_file_dump
function was added to cellblender_mol_viz.py to dump a single viz file.

The function must be called manually (typically inside the Blender
console), and it simply prints the numbers of each molecule type found
in the file.