Removed spaces around "+" in reactants/products for BNGL compatibility.
Earlier versions automatically added spaces around "+" in reactants and
products for better readability. However, BNGL also uses "+" as part of
its syntax, and it does not tolerate the added spaces.

This commit comments out the automatic addition of those spaces. A
better solution might be to parse the reaction strings and decide which
plus signs are part of the reaction syntax and which are part of the
BNGL syntax.