Added operators to build a "stick" figure (actually points) of mols
This change adds to operators to cellblender_molecules.py:

  mol_comp_stick
  mol_comp_nostick

The first one builds a "stick" figure from the currently selected
molecule, and the second one deletes it.

The "stick" figure was originally intended to contain radial lines from
the origin to the locations each component. However, just drawing the
points (not the sticks) looks pretty good so far. It's not clear whether
adding the sticks would be helpful (it may be an option).

Right now, this only works for the "3D Auto" molecule geometry, but it
should be easily extended to all the other molecule generation methods.

The code builds a mesh named:

  mol_comp_stick_mesh

and an object named:

  mol_comp_stick_mesh_shape

These are hard-coded names in this version. These are just a quick way
to show the components in a molecule.