Added a process.wait() after starting the mdlr2mdl.py (race condition)
CellBlender appears to be calling mdlr2mdl.py in two different places:

  cellblender_simulation.py in MCELL_OT_run_simulation_sweep_queue()
  data_model_to_mdl.py in write_mdlr()

The first call was waiting for the process to complete, but the second
one was not. This caused a race condition leading to unpredictable
crashes of MCell when running MDLR models.

The proper solution is to figure out which location should be doing the
actual exporting. However, for now, this commit simply waits after the
call so that there is no race condition (everything is synchronous).