Added a spatial_structure key to molecule data model
Spatial structure may be specified through a number of mechanisms. This
new flag defines the mechanism used for a particular molecule.

The values and display names for this flag are currently defined as:

  'None'    "Coincident"
  '2DAuto'  "2D Auto"
  '3DAuto'  "3D Auto"
  'XYZ'     "XYZ Specified"
  'XYZA'    "XYZ,A Specified"
  'XYZVA'   "XYZ,V,A Specified"

Older models (presumably without spatial structure) are converted to
"None" when upgraded.

Note that the 2 "Auto" modes have another Blender property named:

  component_distance

This is used when computing the "automatic" values for the components.
It's not clear whether this should be explicitly stored in the data
model or if the resulting locations (already in the model) are
sufficient to recreate that value. For now, that value is NOT in the
data model.