Changed data_model_to_mdl.write_mdlr to include extra mcell directory
The "final_mcell_path" was previously finding mcell directly in the
extensions directory. This commit adds the extra "mcell" directory
level to "final_mcell_path".

Note that this requires changes to mdlr2mdl.py as follows:

  diff --git a/mcell/mdlr2mdl.py b/mcell/mdlr2mdl.py
  index d9bf5ce..56f2cac 100755
  --- a/mcell/mdlr2mdl.py
  +++ b/mcell/mdlr2mdl.py
  @@ -36,8 +36,8 @@ class MDLR2MDL(object):
       def __init__(self, configpath):
           self.config = {}
           self.config['bionetgen'] = os.path.join(configpath,'bng2','BNG2.pl')
  -        self.config['mcell'] = os.path.join(configpath,'mcell')
  -        self.config['libpath'] = os.path.join(configpath,'lib')
  +        self.config['mcell'] = os.path.join(configpath,'mcell','mcell')
  +        self.config['libpath'] = os.path.join(configpath,'mcell','lib')
           if (sys.platform == 'linux') or (sys.platform == 'linux2'):
               extension = "so"
           elif (sys.platform == 'darwin'):
  @@ -158,7 +158,7 @@ class MDLR2MDL(object):

   if __name__ == "__main__":
  -    mdlr2mdl = MDLR2MDL(os.path.join(get_script_path()))
  +    mdlr2mdl = MDLR2MDL(os.path.dirname(get_script_path()))

       parser = define_console()
       namespace = parser.parse_args()

These changes cannot be made here because this branch is based on the
development branch which does not include mdlr2mdl.py. However, the
changes in this branch will work for non-MCellR models, and that's why
these changes were made in the development branch and not in the version
that supports MCellR.