Added CellBlender support for optional arguments to MCell
MCell can take a number of arguments. For example, the recently added
"-d" option allows dumping of certain internal MCell data for use outside
of MCell. In this case, the molecule/component structure can be dumped
and used in other programs.

This CellBlender commit adds a new field named "Command Options" in the
"Advanced Options" portion of the "Run Simulation" panel (at the bottom).

As an example, this field can be used to produce MCell's help message
by specifying the "--help" (or "-h") options. As another exmaple, the
"-d 20" option can be specified to set the "dump level" in MCell to show
the molecule structures that it builds (mentioned above).