Added cellblender_molmaker.py to help define and show structure.
This is adapted to CellBlender from a previously stand-alone program
known as "MolMaker.py". The current version takes its input from a text
file stored in Blender and builds a molecule from that structure. This
text file input method may be preserved along with adding Blender panels
to provide that input in a more natural manner.

This is being added to the "component_location" branch.