The Molecule Structure Tool is testable in CellBlender but has problems
This version allows definition of a complex using the molecules already
defined in CellBlender. So if an "L" and an "R" molecule have been
defined, then an L.R.L.R.R molecule can be created. This will build a
panel allowing the definition of the connections between them (which
components bind to which components and at what angle).

The angle doesn't seem to work in all cases, and there may be some other
problems that need to be worked out. This provides an easier way to
build these molecules than the text inputs of earlier versions.