This version uses a Rotation Reference of -1 to indicate a key.
The overloading of "Rotation Reference" to indicate a rotation key
(rather than a normal component) will probably create confusion. This
was done because the panel space for components is getting so crowded.
However, it might be possible to add a check box to the index column so
that it can provide the flag and the index in the same column.