Added the "is_key" member to the MCellMolComponentProperty class
We are beginning to address the rotation issue when binding two
molecules to form a complex. The rotations are defined with respect to
reference locations in each molecule. Often, these reference locations
can be normal components in the molecule. However, other times, this is
either impossible (molecules with either one component or with all
components along a line) or inconvenient. In those cases, it's helpful
to be able to define special "components" to act as rotation keys. These
special "key" components should not appear in the BNGL or in normal
renderings of the molecules. For these reasons, it's helpful to have a
flag to specify when a component is acting only as a key. The "is_key"
flag is intended for that purpose.