Added the rotation indexing to the Molecule Structure Tool
The Molecule Structure Tool allows constructing a 3D model from a
CellBlender component location specification and a bond/angle mapping.

First, all of the molecules should be defined in CellBlender using the
XYZ,AngleRef specification. This requires an (x,y,z) coordinate for each
component along with an index of the component that provides the angle
reference (any point NOT on the line through the origin and this
component). Components that are never bound may use -1 as their
reference index. In the future, it might be better to pick default
reference indexes for each component because binding to a component
without a valid reference will cause an error.

Second, the Molecule Structure Tool is used to build the connectivity
between molecule/component pairs. This is started by defining a simple
"BNGL-like" molecule complex definition using just the molecule names.
For example "L.R" or "L.R.L.R.L" would work. This gives the tool enough
information to lay out these molecules and all of their components to be
connected with unique index numbers (shown on the left of the panel).

The connections are made by referencing the index numbers as the bond
index for the binding component. In other words, each component must
reference its binding partner. Rotation angles can also be specified on
either (or both) of the binding components.

The current code builds complex molecules from a text description that
looks like the following:

  [0] = T (m) at (0.0,0.0,0.0) with peers [1,2,3,4,5,6]
  [1] = t (c) at (0.0,0.0,0.0) with peers [0,8] <5,0.0>
  [2] = t (c) at (0.0,0.0,0.0) with peers [0]
  [3] = t (c) at (0.0,0.0,0.0) with peers [0]
  [4] = t (c) at (0.0,0.0,0.0) with peers [0] <2,0.0>
  [5] = t (c) at (0.0,0.0,0.0) with peers [0]
  [6] = t (c) at (0.0,0.0,0.0) with peers [0]
  [7] = T (m) at (0.0,0.0,0.0) with peers [8,9,10,11,12,13]
  [8] = t (c) at (0.0,0.0,0.0) with peers [7,1] <12,0.5>
  [9] = t (c) at (0.0,0.0,0.0) with peers [7]
  [10] = t (c) at (0.0,0.0,0.0) with peers [7]
  [11] = t (c) at (0.0,0.0,0.0) with peers [7,15] <9,0.5>
  [12] = t (c) at (0.0,0.0,0.0) with peers [7]
  [13] = t (c) at (0.0,0.0,0.0) with peers [7]
  [14] = T (m) at (0.0,0.0,0.0) with peers [15,16,17,18,19,20]
  [15] = t (c) at (0.0,0.0,0.0) with peers [14,11] <19,0.0>
  [16] = t (c) at (0.0,0.0,0.0) with peers [14]
  [17] = t (c) at (0.0,0.0,0.0) with peers [14]
  [18] = t (c) at (0.0,0.0,0.0) with peers [14] <16,0.0>
  [19] = t (c) at (0.0,0.0,0.0) with peers [14]
  [20] = t (c) at (0.0,0.0,0.0) with peers [14]

Descriptions like these can be used directly (via a Blender "text"),
and the tool builds these text descriptions and then passes them to
the molecule parser/builder to actually construct the molecule complex.

This is still very preliminary, but has worked in a few simple tests.