Fixed the component angle referencing
The "molecule/component" (molcomp) array contains both molecules and
components. It is also a sequential list of molecules with each molecule
followed by all of its components.

However, in CellBlender, the components are shown as starting from zero
for each molecule (no accumulation, and no molecule indexes). So the
translation between the CellBlender representation and the "molcomp"
representation needed to account for these differences. The previous
version was producing the wrong component indexes when generating the
"molcomp" representation and some of the angle references were wrong.
This produced some degenerate reference triangles (two of the points
were actually the same). This commit fixes this with proper accounting
of the slot assignments in the "molcomp" array.