Added ability to turn off rotation calculations to help isolate problem.
There are two fairly complex portions of code in:

  bind_molecules_at_components ( ... )

The first is the location calculation, and the second is the axial
rotation. This flag permits turning off the rotation (which is also
helpful when building purely 2D molecules).

At the current time, this flag seems to show that the "bending" problem
is NOT in the axial rotation code.