Updated the older Build 2D and 3D from text to ignore keys in counting.
The addition of keys to the molecule's peer list created an error in the
code that built the 2D and 3D molecules from text. They had assumed that
all non-molecules were components and used that number as a basis for
the component geometry. This change builds a separate list of components
only to use for those calculations.