Fixed the "bending" problem by normalizing the unit vector again.
This version replaced the previous calculation with a quaternion
construction of the rotation matrix. The quaternion version didn't
exhibit the "bending" problem of the previous version. However, after
more detailed exploration, it wasn't the quaternion that solved the
problem. It was a proper normalization of the unit vector used to build
the rotation matrix. Once that was done, both versions appeared to
produce the same result (from visual inspection).

A manual switch was added (if True) to select between the non-quaternion
and the quaternion versions of the calculation. This is currently in the
"non-quaternion" setting.