Added code to handle axis lengths smaller than 1e-30 as a special case.
The cross product is used to generate a rotation axis for aligning
components when binding two molecules. However, when they are already
aligned, the cross product can produce an axis with a length of zero.
When this happens, there is either no rotation needed or the rotation
is 180 degrees (a reversal of direction). These two cases can then be
resolved with the dot product.

This change considers any rotation axis length less than 1e-30 as
falling into the "already aligned" case.