Added support for using CellBlender's molecule colors in MolMaker.
This version has an option (defaulted on) to look up the molecule colors
used by CellBlender for each molecule in the complex. The alternative is
to give them all the same color (defaulted to red).

Since MolMaker is now integrated with CellBlender, it makes sense to use
more of CellBlender's data to be consistent. This version provides that
option.