Fixed 3DAuto problem and disabled several unsupported modes.
Fixed a problem when using keys with the 3DAuto component placement.
The component placement algorithms must be careful to ignore
"components" that are really keys (and not components) when placing
components in space.

Also commented out several of the modes that are not fully supported at
this time to reduce confusion when selecting a mode. The only modes that
should be chosen are the default ("Coincident") mode where there is no
spatial structure and the "XYZ,Angle,Ref" mode which supports choosing
both locations and reference keys for rotation.

The relatively new "2D" and "3D" buttons now support the roles of the
2DAuto and 3DAuto modes. This required adding the "R" (radius) field
beside the 2D and 3D buttons.