Added an option to hide the text interface (defaults to hidden).
The Molecule Structure Tool was originally designed around a text
input file produced by MCell. This text interface is essentially a
list of molecule/component pairs.

The text interface is still useful for viewing the output of MCell
molecules and for interchange of molecule models. However, this is
a less common usage, so it is now being hidden to reduce confusion.