Added support for dynamic angle rotation (and a few debug features)
This version has an option to allow dynamic changing of the various
binding angles. The results are dynamically reflected in the Blender
object named "Mol Object".

This version demonstrated an odd behaviour of rotating part of the
molecule along a different axis when an earlier (different) axis was
also being rotated. The odd thing is that this seemed to happen only
when the earlier rotation was in the negative "z" part of the space.