Converted "build_all_mols" to take a molcomp list rather than text.
Previous versions of build_all_mols accepted a text definition which was
parsed inside build_all_mols. This version places the parsing outside of
that function to speed it up. Note that at this time, the Blender
properties are still being converted to text and then reconverted to a
molcomp list to pass to build_all_mols. But this change opens the door
to a simpler conversion when not starting from text.