Refactored code to isolate all of the text processing.
The text processing is now done in these two functions:

  read_molcomp_data_from_text ( fdata )
  update_molcomp_list_from_defs ( context, molcomp_list, moldef_text ... )

These functions are called directly when reading from a text source.
They are also currently called when CellBlender's data is first
converted to text and then "unconverted" via these functions inside
"build_complex_from_cellblender".

The final step in this refactoring will be to remove the need for any
text translations when working with CellBlender's data. In this case,
the molcomp_list will be built directly from CellBlender's properties.