Added a "Purge by Re-Open" button to clean up the data blocks.
Using the "Dynamic Rotation" option will delete and rebuild the the
object ("Mol Object") with every Bond Angle change. Initially this
happens relatively quickly. However, over time, there seems to be some
kind of accumulation of something (maybe a Blender memory leak).

The following snippet was found on blender.stackexchange.com:

  # save and re-open the file to clean up the data blocks
  bpy.ops.wm.save_as_mainfile(filepath=bpy.data.filepath)
  bpy.ops.wm.open_mainfile(filepath=bpy.data.filepath)

Since this is a somewhat expensive operation, it can't be done for each
angle change. Instead, it's being provided as an operator in this
version. It does seem to make Blender more responsive afterward.