Enabled and used the quaternion rotation versions.
There are two locations in "bind_molecules_at_components" that need a
rotation matrix:

  1. Aligning the axes of two components to be joined
  2. Aligning the keys of two molecules joined along an axis

The matrix is named "R" in both cases, and it can be created either
directly or as a quaternion. They should produce the same matrix in
either case, but there may be subtle differences in edge cases.

The previous code had used the direct rotation matrix, but there has
been an ongoing rotation problem with complexes containing several
molecules. This version enabled the quaternion formulations to see if
they would solve the problem. The change did not have any effect.

Also note that both of these current formulations are not optimized with
respect to computation time. They both compute several values which are
symmetric in their matrices and could be optimized further. This hasn't
been done yet to keep the matrices looking as they would in a formula.