Resolved the ambiguity of acos(dot_prod) using cross product for axial rotation.
The odd axial rotation behavior appears to have arisen from the acos of
the dot product without using the cross product as the axis of rotation.

The code had been using the vector from the variable molecule to its
binding component as the axis of rotation (since it already existed).
The vector was normalized to be "var_rot_unit". However, the rotation
angle was taken from the arc cosine of the dot product. This rotation
angle varies only between 0 and pi. This caused the resulting rotation
angle to reverse itself as observed in the 3D views of the result.

For now, a quick fix was added to compute the cross product and use its
sign to decide if the acos result is correct or if it should be subtracted
from 2*pi to get the continuously varying angle from 0 to 360.

This version appears to be working properly, however, the result was
arrived at through a lot of trial and error (as shown here). Now that
it's working, it would be good to go through the code and carefully
decide the best (and most efficient) way to get that same result.