Changed sense of rotation angle: positive is CCW rotation of peer.
The current formulation includes 2 binding angles - one for each pair of
binding partners. Either or both may be zero.

The sense of the rotation is relative to the "viewpoint" of each
molecule looking through the binding point (component pair) to the
molecule of the binding partner. From this viewpoint, a positive angle
will result in a counter clockwise rotation of the binding partner.

One consequence of this convention is that rotations between binding
partners will add to each other.