Added an "auto keying" operator for automatically setting all keys
It's common to want to set the rotation angle reference key for all
components in a molecule to the same key. It's also common to only have
one such key in a given molecule definition.

This button allows setting of all unset keys (negative index) to the
first key found within a molecule. This will not change any keys that
are already defined with a non-negative index number.