Added an "Average Coincident" option to "fix" miscalculations.
Using the current molecule model, when two components bind, they occupy
the same point in space. In other words, component locations are the
joining points for molecules.

This model works naturally when there are no loops, and all molecules
and components end up wherever their geometry dictates. However, when
there are loops, there is normally no guarantee that the closing points
of a loop are actually coincident in space. For example, if a molecule
were designed as a square (four molecules each binding to each other
with components at 90 degrees), then the geometry would dictate that
the closing points would be coincident. But if the angles were specified
(incorrectly) as 91 degrees rather than 90, then the "closing" components
would not actually join. This could be considered as either a fundamental
error in the model or as a minor round off error. While that decision is
up to the designer, the current tool does provide a simple method to "fix"
any such "round off" errors using the "Average Coincident" option.

The "Average Coincident" option performs one final step after building
the molecule. It sets the location of each partner of a binding pair to
be the average location of the two binding partners. For small rounding
errors, this simply brings the points into exact coincidence. However,
for gross errors (such as forming an equilateral triangle with right
angles), it will drastically change the geometry. The decision to use
this feature (or not) is up to the model's designer.