Restored previously removed molecule geometry modes as "- Unsupported -"
As the spatially structured molecules have evolved, some of the early
prototype modes had been removed from the user options by removing them
from the EnumProperty that's used to display them.

However, any older ".blend" files that had used those now missing modes
would fail to run and produce error messages that were not obvious. The
problem appears to be due to "ID Property" values that don't match the
currently allowed "RNA Property" values for that EnumProperty.

The proper fix is probably to upgrade those data models to use an
appropriate structured molecule model that will be supported in the
future. But since those were mostly experments, the shortcut in this
commit is to simply retain the older EnumProperty values, but draw them
as "- Unsupported -" in the panel.

As a practical matter, the actual property values stored for the
different spatial representations were mostly the same (x,y,z,...),
so it's reasonable to consider them to be subsets of "XYZRef". But
making that determination would require a little more investigation.
At this time, it's more important to focus on the modes that will most
likely be used (Coincident and XYZRef).

Note that the commit that removed the unsupported modes was:

  Commit: eee430218fcc6fefbf345ab012ed96eeff8cab74

That was Oct 23, 2018 at 12:39:54. Any of the CellBlender versions
between that commit and this current commit would not be able to open
".blend" files that contained any of the removed modes.