Don't write "key" components to the MDLR file.
The addition of "key" components (which represent a spatial location
while not being an actual component) created a problem when those
"components" were exported. To get around this quickly, this commit
creates a second list with the "key" components removed. That second
list is then used for all subsequent processing.

This is a quick work-around to verify that the rest of the system still
works. The next step will be to put the "key" locations into the
"rot_x", "rot_y", and "rot_z" fields. This will mean writing them out
for every component rather than using a component "reference".