Export the key locations in the rotation axis "rot_x,y,z" fields.
The spatial syntax for components is:

  {loc=[lx,ly,lz], rot=[rx,ry,rz,a]}

When using components (or keys) as reference locations, the rx,ry,rz
fields will come from another component's lx,ly,lz fields. This commit
exports those values from the referenced component. Note that they are
NOT being normalized at this time. It's not clear whether they should be
normalized before putting into the "rot=[]" syntax or whether MCell
itself should always normalize them anyway. If we allow people to write
their own MDLR, then MCell should probably NOT assume that they'll be
normalized. That's the policy of this commit. That makes a little more
work for MCell, but that work is likely to only be needed as it reads
the MDLR. The same work would have to be done by CellBlender anyway, so
it's not likely to be much of a performance loss.

This commit also changed the name of "fake_m" to the more appropriate
"no_key_m" which describes it as "m" with the keys removed. This isn't
quite true since it only contains the "bngl_component_list" portion of
the actual "m", but it's more descriptive than "fake_m".