Removed names from unsupported molecule structure options.
As mentioned in an earlier commit, there's a problem with removing
EnumProperty options when reading in older .blend files. Attempting
to set these "non-existent" values causes a difficult to understand
error message from Blender. So removing items will break backward
compatibility with older files. It would be nice if they could just
not be drawn in the enumeration, but there's no easy way to do that.
This commit at least groups the useful options together and draws the
unsupported options as a row of dashes.

It would be possible to use the data model upgrade process to map the
older unsupported options and fields into the current model. But that
will take a bit more work and isn't justified at this time since the
final options are still being evaluated. Eventually that will be the
solution, but for now, the line of dashes should let users know not to
choose those options.