Replaced newline (\n) with linesep in cellblender_utils.try_to_import.
The "try_to_import" function tests the availability of Python modules in
an external Python environment by building a small script that simply
attempts to import each of the requested modules. It runs the script in
the requested external Python and converts the returned result into a
pass/fail return code.

Prior to this commit, the small text script was built with newline
characters (\n) separating each line. This version uses os.linesep
instead.

Since the resulting program will be run inside the Python interpreter,
it's not clear if this would be an error or not under operating systems
with differing line endings. But using os.linesep is the proper way in
any case.