Added a loop limit to the object sorting in MDLR processing.
CellBlender's MDLR processing requires sorting objects according to
their physical nesting. The previous algorithm will fail if an object
declares a parent that doesn't exist. This is an error, but it should
not cause CellBlender to enter an infinite loop while exporting. That
was the case with the previous logic. The use of the "any_change" flag
ensures that the while loop will exit if it cannot move any objects from
the "unsorted" list to the "sorted" list (for any reason).