Updated the MolMaker portion of the example to match the model.
The MolMaker panel is currently saved in the Data Model for the user's
convenience. But what's shown in the MolMaker panel doesn't have to
match what's in the actual model. That's normally not a problem, but for
an example, it's nice for them to match. This commit updated the
MolMaker panel to build a short chain of just a few molecules.